Natural sort order to list all command
Running `asdf list all julia` currently does not list Julia versions in natural sort order. Instead, we get something like the following:

```
...
1.0.5
1.10.0-alpha1
1.10.0-beta1
1.10.0-beta2
1.10.0-beta3
1.10.0-rc1
1.10.0-rc2
1.10.0-rc3
1.10.0
1.1.0-rc1
1.1.0-rc2
1.1.0
1.1.1
...
1.9.0-rc3
1.9.0
1.9.1
1.9.2
1.9.3
1.9.4
```

This is clearly less than ideal, especially with the recent release of version 1.10, which gets buried further up into the list where most people won't initially think to look for it (like I did the other day).

This PR adds the `-V` flag to the `sort` command in the `sort_versions` function, we get natural order version sorting. 

However, my understanding is that not all operating systems have a version of `sort` that supports that flag. Therefore, let's treat this PR as the start the process of addressing this issue. I'm sure there is an OS/version agnostic way of getting natural sort, I just don't know what it is yet.